+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v3.99.4
 * Fixed bug with wrong context on session closing
 * Fixed goroutine leak on closing `database/sql` driver
 * "No endpoints" is retriable error now